cleanup basic types
